Rename/rearrange classes from review
